@@ -49,13 +49,12 @@
     "body-parser": "^1.15.2",
     "classnames": "^2.2.5",
     "cors": "^2.8.3",
-    "css-loader": "^0.27.3",
+    "css-loader": "^6.7.1",
     "date-fns": "1.29.0",
     "dayjs": "^1.8.27",
     "dotenv": "^4.0.0",
     "dotenv-webpack": "^1.7.0",
     "ejs": "^2.5.6",
-    "erxes-icon": "^1.0.5",
     "express": "^4.14.0",
     "fast-equals": "^2.0.4",
     "file-loader": "^1.1.11",
@@ -63,7 +62,7 @@
     "graphql-tag": "^2.6.1",
     "graphql-ws": "^5.5.5",
     "i18n-react": "^0.6.4",
-    "json-loader": "^0.5.4",
+    "json-loader": "^0.5.7",
     "node-sass": "^4.14.1",
     "qrcode": "^1.5.0",
     "react": "^16.11.0",
@@ -73,12 +72,12 @@
     "react-transition-group": "^2.4.0",
     "sass-loader": "^6.0.6",
     "source-map-loader": "^0.2.3",
-    "style-loader": "^0.16.1",
+    "style-loader": "^3.3.1",
     "subscriptions-transport-ws": "^0.9.14",
-    "uglifyjs-webpack-plugin": "1.2.4",
-    "url-loader": "^1.0.1",
-    "webpack": "4.16.3",
-    "webpack-cli": "^3.1.0",
+    "terser-webpack-plugin": "^5.3.1",
+    "url-loader": "^4.1.1",
+    "webpack": "^5.57.1",
+    "webpack-cli": "^4.9.0",
     "webpack-merge": "^4.1.3",
     "xss": "^1.0.3"
   },
