@@ -1,8 +1,9 @@
 import * as React from "react";
-import { App as DumbApp } from "../components";
+import DumbApp from "../components/App";
 import { connection } from "../connection";
 import { AppConsumer, AppProvider } from "./AppContext";
 import { postMessage, saveBrowserInfo } from "./utils";
+import "../sass/style.scss";
 
 type Props = {
   loadType: string;
