@@ -1,5 +1,5 @@
 import * as React from "react";
-import { Category } from "../containers";
+import Category from "../containers/Category";
 import { IKbTopic } from "../types";
 
 type Props = {
