@@ -5,7 +5,8 @@ import * as xss from 'xss';
 import { defaultAvatar } from '../../icons/Icons';
 import { IUser } from '../../types';
 import { __, readFile, urlify } from '../../utils';
-import { Attachment, User } from '../components/common';
+import Attachment from '../components/common/Attachment';
+import User from '../components/common/User';
 import { MESSAGE_TYPES } from '../containers/AppContext';
 import { IAttachment, IMessengerAppData, IVideoCallData } from '../types';
 import VideoCallMessage from './VideoCallMessage';
