@@ -1,8 +1,40 @@
 import * as React from "react";
 import { IUser } from "../../types";
-import { AccquireInformation, ConversationDetail, Home, ConversationList} from "../containers";
-import { ArticleDetail, CategoryDetail } from "../containers/faq";
-import WebsiteAppDetail from "../containers/websiteApp/WebsiteAppDetail";
+import asyncComponent from "../../AsyncComponent";
+
+const ConversationDetail = asyncComponent(() => 
+  import(/* webpackChunkName: "MessengerConversationDetail" */ "../containers/ConversationDetail")
+);
+
+const AccquireInformation = asyncComponent(() => 
+  import(/* webpackChunkName: "MessengerAcquireInformation" */ "../containers/AccquireInformation")
+);
+
+const Home = asyncComponent(() => 
+  import(/* webpackChunkName: "MessengerHome" */ "../containers/Home")
+);
+
+const ConversationList = asyncComponent(() => 
+  import(/* webpackChunkName: "MessengerConversationList" */ "../containers/ConversationList")
+);
+
+const ArticleDetail = asyncComponent(() =>
+  import(
+    /* webpackChunkName: "MessengerArticleDetail" */ '../containers/faq/ArticleDetail'
+  )
+);
+
+const CategoryDetail = asyncComponent(() =>
+  import(
+    /* webpackChunkName: "MessengerCategoryDetail" */ '../containers/faq/CategoryDetail'
+  )
+);
+
+const WebsiteAppDetail = asyncComponent(() =>
+  import(
+    /* webpackChunkName: "MessengerWebsiteDetail" */ '../containers/websiteApp/WebsiteAppDetail'
+  )
+);
 
 type Props = {
   activeRoute: string | "";
