@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { MessagesList } from '../components';
+import MessagesList from '../components/MessagesList';
 import { IMessage } from '../types';
 import { AppConsumer } from './AppContext';
 
