@@ -19,6 +19,7 @@ module.exports = {
   output: {
     path: path.join(__dirname, 'static'),
     filename: '[name].bundle.js',
+    chunkFilename: '[name].[contenthash].js'
   },
 
   plugins: [
@@ -30,22 +31,29 @@ module.exports = {
       {
         test: /\.(t|j)sx?$/,
         loader: 'awesome-typescript-loader',
-        exclude: /node_modules/,
+        exclude: /node_modules/
       },
       // addition - add source-map support
       {
-        enforce: "pre",
+        enforce: 'pre',
         test: /\.js$/,
-        loader: "source-map-loader",
+        loader: 'source-map-loader',
         exclude: [/node_modules/]
       },
       {
         test: /\.css$/,
-        use: ['style-loader', 'css-loader'],
+        use: ['style-loader', 'css-loader']
       },
       {
-        test: /\.(png|woff|woff2|eot|ttf|svg)$/,
-        loader: 'url-loader?limit=100000',
+        test: /\.(png|woff|woff2|eot|ttf)$/,
+        use: [
+          {
+            loader: 'url-loader',
+            options: {
+              limit: 100000
+            }
+          }
+        ]
       },
       {
         test: /\.scss$/,
@@ -54,9 +62,17 @@ module.exports = {
       },
       {
         test: /\.(eot|woff|woff2|ttf|svg|png|jpg|gif)$/,
-        loader: 'url-loader?limit=30000&name=[name]-[hash].[ext]'
+        use: [
+          {
+            loader: 'url-loader',
+            options: {
+              limit: 30000,
+              name: '[name]-[hash].[ext]'
+            }
+          }
+        ]
       }
-    ],
+    ]
   },
 
   // addition - add source-map support
