@@ -1,5 +1,5 @@
 const webpack = require('webpack');
-const UglifyJSPlugin = require('uglifyjs-webpack-plugin');
+const TerserPlugin = require("terser-webpack-plugin");
 const Merge = require('webpack-merge');
 const CommonConfig = require('./webpack.config');
 
@@ -10,22 +10,23 @@ module.exports = Merge(CommonConfig, {
       debug: false,
     }),
     new webpack.DefinePlugin({
-      'process.env.NODE_ENV': '"production"',
+      'process.env.NODE_ENV': '"production"'
     }),
-    new UglifyJSPlugin({
-      test: /\.js($|\?)/i,
-      sourceMap: true,
-      uglifyOptions: {
-        mangle: {
-          keep_fnames: true,
+    new TerserPlugin({
+      terserOptions: {
+        ie8: false,
+        ecma: 8,
+        mangle: true,
+        output: {
+          comments: false,
+          beautify: false,
         },
         compress: {
-          warnings: false,
-        },
-        output: {
-          beautify: false,
+          pure_getters: true,
+          unsafe: true,
+          unsafe_comps: true,
         },
       },
     }),
-  ],
-})
+  ]
+});
